Refactoring Mappers and Exceptions Packages
