+++ src/main/java/com/unsa/suppliers/infrastructure/controllers/RoleController.java
@@ -6,7 +6,7 @@ import com.unsa.suppliers.domain.dtos.roles.RoleResponse;
 import com.unsa.suppliers.domain.entities.RoleEntity;
 import com.unsa.suppliers.domain.exceptions.roles.RoleDuplicatedException;
 import com.unsa.suppliers.domain.exceptions.roles.RoleNotFoundException;
-import com.unsa.suppliers.domain.mappers.RoleMapper;
+import com.unsa.suppliers.application.mappers.RoleMapper;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
